requirement.txt
# The following contains only Python package dependencies

Pillow==10.2.0
ultralytics==8.1.0
Cython>=0.28.5
Django>=1.11.7
thulac>=0.1.2
py2neo==4.1.0
pyfasttext==0.4.5
pinyin>=0.4.0
pymongo>=3.6.1
